windows: avoid `argv[0]` for MSVC Racket build
WSL2 launches a Windows executable with only the executable name, and
not the path used to launch the executable. That breaks finding the
"collects" directory, at least for "Racket.exe" as compiled with MSVC.
In other dbuild modeswords, `argv[0]` is already replaced with the
path as found via `GetModuleFileNameW`; that avoids the problem, and
it's better to be consistent.

Closes #3867