raco setup: adjust layering support for libraries
Refrain from installing a library when it exists in the search path
only if the one in the search path matches the library that would be
installed. This change is intended to make layering support compatible
with setting a search path to find extra libraries, where the
extra-libraries directory has libraries with the same name as Racket
libraries, but they're not meant to replace Racket libraries (e.g.,
libcairo2.dylib installed via Homebrew on Mac OS).